requirements.txt
streamlit
google-generativeai
pandas
plotly
python-dotenv
watchdog
psycopg2-binary
